Refactoring weekend checks, assorted upgrades.
Fixed a bug where, if yesterday was Sunday, the range in the Timesheets query would be invalid. This was a result of yesterday getting set to last Friday, but the firstDayInWeek date object was set to today (Monday). I abandoned the notion that the checkTime function would not run on weekends since it was making the code more complex than it needed to be. Now getDates doesn't oddly return a false values if yesterday met certain conditions; instead, it always returns a date and checkTime always completes, though it won't send emails on the weekends.

There are various other upgrades included, such as better comments and pseudo-constant variable names. The spreadsheet that is produced also provides more information, as does the email users will receive if they don't track enough time.