Making timecop convert PTO days to 7 hours.
Since we need to track 8 hours for PTO and Holidays, I need to convert those time entries to 7 hours. This lets TimeCop make accurate decisions regarding when to email users. For example, if a user had an 8 hour day earlier in the week, that may have allowed them to track a 6 hour day later and not receive an email reminding them to track their time.